+++ src/widgets/layout/configurator.jsx
@@ -57,7 +57,7 @@ export function Configurator() {
 
   React.useEffect(() => {
     const stars = fetch(
-      "https://api.github.com/repos/creativetimofficial/material-tailwind-dashboard-react"
+      "https://api.github.com/repos/creativetimofficial/vehicle-dashboard"
     )
       .then((response) => response.json())
       .then((data) => setStars(formatNumber(data.stargazers_count, 1)));
@@ -148,7 +148,7 @@ export function Configurator() {
           <hr />
           <div className="my-8 flex flex-col gap-4">
             <a
-              href="https://www.creative-tim.com/product/material-tailwind-dashboard-react?rel=mtdr"
+              href="https://www.creative-tim.com/product/vehicle-dashboard?rel=mtdr"
               target="_black"
             >
               <Button variant="gradient" fullWidth>
@@ -174,7 +174,7 @@ export function Configurator() {
           </div>
           <a
             className="mx-auto flex items-center justify-center gap-2"
-            href="https://github.com/creativetimofficial/material-tailwind-dashboard-react"
+            href="https://github.com/creativetimofficial/vehicle-dashboard"
             target="_blank"
             rel="noreferrer"
           >