+++ src/widgets/layout/navbar.jsx
@@ -83,10 +83,10 @@ export function Navbar({ brandName, routes, action }) {
 }
 
 Navbar.defaultProps = {
-  brandName: "Material Tailwind React",
+  brandName: "VECV",
   action: (
     <a
-      href="https://www.creative-tim.com/product/material-tailwind-dashboard-react"
+      href="https://www.creative-tim.com/product/vehicle-dashboard"
       target="_blank"
     >
       <Button variant="gradient" size="sm" fullWidth>